Fix the test that never compiled on CI
CI has failed all 14 times it has ever run, always at the same place: an array
literal whose element type has to be inferred.

  CircularClock.averageMinutes([23 * 60 + 50, 10])

averageMinutes takes [Double]. Xcode 26.6 types the arithmetic as Double and
compiles it; the runner's Xcode 26.3 types it as Int and refuses to convert.
So the whole suite failed to build on every push while passing locally, which
is the worst shape a CI failure can take: it looks like noise, and it trained
us to ignore the mail.

An explicit Double() removes the inference entirely, so both compilers agree.

Also corrects the workflow comment, which promised 52 unit and 7 UI tests when
the suite is 206 and 11.

+++ Health TrackerTests/Health_TrackerTests.swift
@@ -79,7 +79,11 @@ struct SleepSessionAssemblerTests {
 
 struct RecapMathTests {
     @Test func circularAverageHandlesMidnight() throws {
-        let average = try #require(CircularClock.averageMinutes([23 * 60 + 50, 10]))
+        // 23:50 and 00:10. The Double() is load-bearing: inside an array
+        // literal the element type has to be inferred, and Xcode 26.3 types
+        // `23 * 60 + 50` as Int and then refuses to convert it. 26.6 resolves
+        // it, so this compiled locally and failed only on CI.
+        let average = try #require(CircularClock.averageMinutes([Double(23 * 60 + 50), 10]))
         #expect(average < 1 || average > 1_439)
     }
